==> analysis/4_Analysis.Rmd <==
---
title: "4_Analysis"
author: "Catherine Arnett"
date: "11/29/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lme4)
load("~/GitHub/LTG-Language_Typology_Geography/clean-data/final_data.RData")

```

# Replicating Everett's findings

First we will attempt to replicate the findings and claims from Caleb Everett's papers (CITE).

## Replicate Ejective-Elevation Correlation

Everett (2013) claims that ejective phonemes are more likely to occur in areas of high elevation (over 1500m). In our model, which looks only at whether elevation (in meters) correlates with having ejectives or not, we found that there was an extremely small but highly significant effect ($\beta$ < 0.001, p < 0.001).

Additionally, we ran another analysis where we included languages having ejectives (in addition to other glottolized features) or not. We again found that there was small but significant effect ($\beta$ < 0.001, p < 0.05).

```{r}
final.data$glot_features_new = final.data$glot_features

final.data$glot_features_new <- ifelse(final.data$glot_features_new == '2 - Ejectives only', 1,
                         ifelse(final.data$glot_features_new == '1 - No glottalized consonants', 0,
                                ifelse(final.data$glot_features_new == '5 - Ejectives and implosives', 1,
                                       ifelse(final.data$glot_features_new == '8 - Ejectives, implosives, and glottalized resonants', 1,
                                              ifelse(final.data$glot_features_new == '6 - Ejectives and glottalized resonants', 1,
                                0)))))



ejectives.elevation <- lm(data = final.data, glot_features_new ~ elev_m)
summary(ejectives.elevation)
anova(ejectives.elevation)

final.data$glot_features_everett = final.data$glot_features
final.data$glot_features_everett <- ifelse(final.data$glot_features_everett == '2 - Ejectives only', 1,
                         ifelse(final.data$glot_features_everett == '1 - No glottalized consonants', 0,
                                ifelse(final.data$glot_features_everett == '8 - Ejectives, implosives, and glottalized resonants', 1,
                                       ifelse(final.data$glot_features_everett == '	7 - Implosives and glottalized resonants', 0,
                                              ifelse(final.data$glot_features_everett == '6 - Ejectives and glottalized resonants', 1,
                                                     ifelse(final.data$glot_features_everett == '	5 - Ejectives and implosives', 1,
                                                            ifelse(final.data$glot_features_everett == '4 - Glottalized resonants only', 0,
                                                                   ifelse(final.data$glot_features_everett == '	3 - Implosives only', 0,
                                1))))))))


ejectives.elevation2 <- lm(data = final.data, glot_features_everett ~ elev_m)
summary(ejectives.elevation2)
```
If we include another predictor, such as language family (`stock`), the effect disappears (SumSq = 0.044, p = 0.177).

```{r}
ejectives.elevation.mult <- lm(data = final.data, glot_features_new ~ stock + elev_m)
anova(ejectives.elevation.mult)
```
If we recode elevation into binary categories `low` and `high`, where high corresponds to any elevation at or above 1500m, there is a larger effect when you don't include language family ($\beta$ = -0.190, p < 0.001) and also when language family is included (SumSq = 0.025, p = 0.4).

```{r}
final.data$elev_binary <- final.data$elev_m

final.data$elev_binary<- ifelse(final.data$elev_binary >= 1500, 'high',
                         ifelse(final.data$elev_binary < 1500, 'low',
                                'low'))

final.data <- final.data %>%
  mutate(elev_binary = as.character(elev_binary))

ejectives.elevation2 <- lm(data = final.data, glot_features_new ~ elev_binary)
summary(ejectives.elevation2)

ejectives.elevation3 <- lm(data = final.data, glot_features_new ~ stock +elev_binary)
anova(ejectives.elevation3)
```
## Replicate C-V Ratio-Latitude Correlation

Everett (2017) makes the connection between Consonant-Vowel Ratio and humidity. (CITE PAPER) connects high C-V ratio and distance from equator. 

There is a correlation between distance from the equator (absolute value of latitude) and a high CV ratio ($\beta$ = 0.018, p < 0.001). However, when you include language family into the model, the effect is no longer significant (SumSq = 0.04, p = 0.812).

```{r}
final.data$cv_ratio_num <- final.data$cv_ratio 

final.data$cv_ratio_num <- ifelse(final.data$cv_ratio_num == '1 - Low', 1,
                              ifelse(final.data$cv_ratio_num == '2 - Moderately low', 2,
                                 ifelse(final.data$cv_ratio_num == '3 - Average', 3,
                                         ifelse(final.data$cv_ratio_num == '4 - Moderately high', 4,
                                                 ifelse(final.data$cv_ratio_num == '5 - High', 5,
                                                          1)))))

final.data <- final.data %>%
  mutate(cv_ratio_num = as.numeric(cv_ratio_num))

final.data$latitude_abs <- final.data$Latitude

final.data <- final.data %>%
  mutate(latitude_abs = abs(latitude_abs))

CV.latitude <- lm(data = final.data, cv_ratio_num ~ latitude_abs)
summary(CV.latitude)

CV.latitude.fam <- lm(data = final.data, cv_ratio_num ~ stock + latitude_abs)
anova(CV.latitude.fam)
```

## Replicate Tone-Humidity Correlation

Everett et al. (2016) claim there is a correlation between low humdity and not having tone, which leads to the implication that tonal languages are more likely to occur in high humidity regions. We also found a significant correlation between tone and humidity ($\beta$ = 18.85, p < 0.01). **TODO: Language family effect explanation here**

```{r}
final.data$tone_num <- final.data$tones

final.data$tone_num <- ifelse(final.data$tone_num  == '1 - No tones', 1,
                              ifelse(final.data$tone_num  == '2 - Simple tone system', 2,
                                 ifelse(final.data$tone_num  == '3 - Complex tone system', 3,
                                                          1)))
final.data <- final.data %>%
  mutate(tone_num = as.numeric(tone_num))

tone.humidity <- lm(data = final.data, tone_num ~ mean_hum)
summary(tone.humidity)

tone.humidity2 <- lm(data = final.data, tone_num ~ stock + mean_hum)
anova(tone.humidity2)
```
When the tone variable was recoded as a binary factor (whether the langauge had any tone system or not), there is still a (less) significant effect of humidity on tone ($\beta$ = 12.44, p < 0.05).

```{r}
#tones/no tones
final.data$tone_bin <- final.data$tone_num

final.data$tone_bin <- ifelse(final.data$tone_bin  == 1, 0,
                              ifelse(final.data$tone_bin == 2, 1,
                                 ifelse(final.data$tone_bin == 3, 1,
                                                          0)))

final.data$tone_bin <- as.numeric(final.data$tone_bin)


tone.humidity2 <- lm(data = final.data, tone_bin ~ mean_hum)
summary(tone.humidity2)
```
# FINAL MODELS REAL FINAL FOR REAL

### CV Ratio

Latitude is not significant ($\chi$^2^(1) = 0.5874, p = 0.443). 

```{r}
# cv.lat.model <- lmer(data = final.data, cv_ratio_num ~ latitude_abs + (1 | stock))
# summary(cv.lat.model)

cv.hum.model <- lm(data = final.data, cv_ratio_num ~ mean_hum + stock)
anova(cv.hum.model)

cv.hum.model.null <- lm(data = final.data, cv_ratio_num ~ mean_hum)
anova(cv.hum.model.null)
```

### Ejectives

I changed this to logistic (glmer).

```{r}
ejectives.full.model <- glmer(data = final.data, glot_features_new ~ elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)
summary(ejectives.full.model)

ejectives.reduced.model <- glmer(data = final.data, glot_features_new ~ mean_hum + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)
anova(ejectives.reduced.model, ejectives.full.model)
```
### Tone

In a model comparison, humidity does seem to be a significant factor on tone ($\chi$^2^(1) = 2074.8, p \< 0.001).

I changed this to logistic (glmer).

```{r}
tone.hum.model <- glmer(data = final.data, tone_bin ~ mean_hum + (1 | stock), family = binomial)
summary(tone.hum.model)

#tone.full.model <- glmer(data = final.data, tone_bin ~ elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)

#tone.reduced.model <- glmer(data = final.data, tone_bin ~ elev_binary + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)
#anova(tone.reduced.model, tone.full.model)
```


# IGNORE BELOW HERE                     

# Try the other combinations of environmental factors and linguistic features

Next, we will look to see if there are any effects of the other environmental factors on the different language features. 

## Ejectives 

### Humidity

There is a strong correlation between ejectives and humidity ($\beta$ = -20.237, p < 0.001), which has a larger effect size than the correlation between ejectives and altitude.

```{r}
ejectives.humdity <- lm(data = final.data, glot_features_new ~ mean_hum)
summary(ejectives.humdity)
```

### Latidude

There is also an effect of latitude, which again is larger than the effect of altitude ($\beta$ = 0.006, p < 0.001). 

```{r}
ejectives.latitude <- lm(data = final.data, glot_features_new ~ latitude_abs)
summary(ejectives.latitude)
```

## C-V Ratio

## Humidity

There is also an effect of humidity on CV ratio ($\beta$ = -82.04, p  < 0.001). It also accounts for more of the variance (R^2^ = 0.141) than the model with latitude (R^2^ = 0.07).

```{r}
cv.humdity <- lm(data = final.data, cv_ratio_num ~ mean_hum)
summary(cv.humdity)
```

## Temperature (vs C-V Ratio) 

There is also an effect of temperature on CV ratio ($\beta$ = -0.024, p  < 0.001). It also accounts for more of the variance (R^2^ = 0.039) than the model with latitude (R^2^ = 0.07), but not the model with humidity (R^2^ = 0.14).

```{r}
final.data$temp_celsius <- (final.data$mean_temp - 273.15) #From Kelvin to Celsius

cv.temperature <- lm(data=final.data, cv_ratio_num~temp_celsius)
summary(cv.temperature)

#hist(final.data$temp_celsius) #there are negative values, so we cannot log-transform?

```
## Altitude

There did seem to be an effect of altitude on CV ratio, but the effect is very small ($\beta$ < 0.001 , p < 0.001).

```{r}
cv.elev <- lm(data = final.data, cv_ratio_num ~ elev_m)
summary(cv.elev)
```

## Tone

Finally, there is no effect of altitude ($\beta$ < 0.001, p = 0.686) but a highly significant effect of latitude.

## Altitude

```{r}
tone.elev <- lm(data = final.data, tone_bin ~ elev_m)
summary(tone.elev)
```

## Latitude

```{r}
tone.lat <- lm(data = final.data, tone_bin ~ latitude_abs)
summary(tone.lat)
```

# Finding alternative correlations

Some factors, which might help explain this apparent correlation between climate and language structure are: a) community size, b) GDP, and c) linguistic diversity of a country or region. We will focus on linguistic diversity.


## Linguistic diversity

We're testing whether languages spoken in countries with high linguistic diversity are more likely to have these features. 

```{r}
#create a new diversity score 

final.data <- final.data %>%
  mutate(diversity.new = established.langs/total.speakers)
```

### Ejectives

There is a significant correlation between high linguistic diversity and having ejectives ($\beta$ = -690.252, p = 0.012).

```{r}
ejectives.diversity <- lm(data = final.data, glot_features_new ~ diversity.new)
summary(ejectives.diversity)
```
### CV Ratio

There is a significant correlation between high CV ratio and high linguistic diversity, although the effect size is extremely small ($\beta$ < 0.001, p < 0.001). If we include language family into the model, there is still no effect (SumSq = 0.03, p < 0.833).

```{r}
cv.diversity <- lm(data = final.data, cv_ratio_num ~ diversity.new)
summary(cv.diversity)

cv.diversity2 <- lm(data = final.data, cv_ratio_num ~ stock + diversity.new)
anova(cv.diversity2)
```

### Tone

There is no significant correlation between linguistic diversity and tone ($\beta$ < 0.001, p = 0.731). 

```{r}
tone.diversity <- lm(data = final.data, tone_num ~ diversity.new)
summary(tone.diversity)

```

# Multiple Regression

## GDP + linguistic diversity

First, it is worth noting that there is a correlation between linguistic diversity and GDP in 2018 (t(353) = -2.959, p = 0.003).

In a model which predicts ejectives from linguistic diversity, there is a marginally significant and very small effect of linguistic diversity ($\beta$ < 0.001, p = 0.084) but not of GDP ($\beta$ < 0.001, p = 0.148). 

```{r}

cor.test(final.data$`2018`, final.data$diversity.new)


ejectives.diversity.GDP <- lm(data = final.data, glot_features_new ~ diversity.new + `2018`)
summary(ejectives.diversity.GDP)
```

##Model Comparison

1) feature ~ environmental + GDP/diversity // compare without environmental 

2) feature ~ GDP/diversity // feature ~ environmental --> compare R2 etc

If we want to compare 

### Ejectives 

In the full model, `mean_hum` (SumSq = 3.430, p \< 0.001) significantly predict having ejectives. In a model comparison, a full model is better than the full model without the elevation (F(1) = 8.543, p = 0.003). 

```{r}
ejectives.all <- lm(data = final.data, glot_features_new ~ stock + elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new)
anova(ejectives.all)

#model without elevation
ejectives.reduced <- lm(data = final.data, glot_features_new ~ stock + mean_hum + mean_temp + latitude_abs + diversity.new )
anova(ejectives.reduced, ejectives.all)
```
### Tone

In the full model, only latitude is significant (F(1) = 4.648, p = 0.032), not humidity. In the model comparison, humidity is not significant.

```{r}
tone.full <- lm(data = final.data, tone_num ~ stock + elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new)
anova(tone.full)

tone.reduced <- lm(data = final.data, tone_num ~ stock + elev_binary + mean_temp + latitude_abs + diversity.new)
anova(tone.reduced, tone.full)

```

### C-V Ratio

In the full model, latitude is a significant factor (F(1) =  6.44, p = 0.012), but in the model comparison, the model is not significant (F(1) = 1.152, p = 0.284). 

```{r}
cv.full <- lm(data = final.data, cv_ratio_num ~ stock + elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new)
anova(cv.full)

cv.reduced <- lm(data = final.data, cv_ratio_num ~ stock + elev_binary + mean_hum + mean_temp + diversity.new)
anova(cv.reduced, cv.full)

```

## Using lme4

### Ejectives

<<<<<<< Updated upstream
# Next check for correlation between environmental features and linguistic diversity


## Linguistic Diversity

There is a significant correlation between linguistic diversity and each altitude, elevation, and humidity. 


### Altitude

```{r}
elev.diversity <- lm(data = final.data, elev_m ~ diversity.new)
summary(elev.diversity)
```

### Latitude 

```{r}
latitude.diversity <- lm(data = final.data, latitude_abs ~ diversity.new)
summary(latitude.diversity)
```

### Humidity
```{r}
hum.diversity <- lm(data = final.data, mean_hum ~ diversity.new)
summary(hum.diversity)
```



#Check for correlation between language features and community size (Ebru)
```{r}

#hist(final.data$total.speakers) 
final.data$total.speakers.log10 <- log10(final.data$total.speakers)
#hist(final.data$total.speakers.log10)

```

##Tone
```{r}
tone.communitysize <- lm(data = final.data, tone_bin ~ total.speakers.log10)
summary(tone.communitysize)


```

##C-V ratio
```{r}
cv.communitysize <- lm(data = final.data, cv_ratio_num ~ total.speakers.log10)
summary(cv.communitysize)

#without taking into account language family or diversity
cv.full.model.less <- lm(data = final.data, cv_ratio_num ~  elev_binary + mean_hum + mean_temp + latitude_abs)
cv.null.model.less <- lm(data = final.data, cv_ratio_num ~  elev_binary + mean_hum + mean_temp)
anova(cv.null.model.less, cv.full.model.less)
```

##Ejectives
```{r}
ejectives.communitysize <- lm(data = final.data, glot_features_new ~ total.speakers.log10)
summary(ejectives.communitysize)

ejectives.communitysize2 <- lm(data = final.data, glot_features_everett ~ total.speakers.log10)
summary(ejectives.communitysize2)
```
=======
In an `lmer` model, if we use stock as the random effect, there is still no effect of elevation on ejectives ($\chi$^2^(1) = 1.588, p = 0.208).


## (G)LMER Models

I changed this to logistic (glmer).

```{r}
final.data <- final.data %>%
  mutate(macroarea = as_factor(macroarea)) 

ejectives.elev.model1 <- glmer(data = final.data, glot_features_new ~ elev_m + (1 | stock) + (1 | macroarea), family = binomial, control=glmerControl(optimizer="bobyqa")) #+ (1 | macroarea)

ejectives.elev.model2 <- glmer(data = final.data, glot_features_new ~ log10(elev_m) + (1 | stock) + (1 | macroarea), family = binomial, control=glmerControl(optimizer="bobyqa"))

ejectives.elev.model3 <- glmer(data = final.data, glot_features_everett ~ elev_m + (1 | stock) + (1 | macroarea), family = binomial, control=glmerControl(optimizer="bobyqa"))

ejectives.elev.model4 <- glmer(data = final.data, glot_features_everett ~ log10(elev_m) + (1 | stock) + (1 | macroarea), family = binomial, control=glmerControl(optimizer="bobyqa"))

ejectives.elev.model5 <- glmer(data = final.data, glot_features_new ~ elev_binary + (1 | stock) + (1 | macroarea), family = binomial, control=glmerControl(optimizer="bobyqa")) #+ (1 | macroarea)


ejectives.null.model <- glmer(data = final.data, glot_features_new ~ (1 | stock) + (1 | macroarea), family = binomial, control=glmerControl(optimizer="bobyqa")) #+ (1 | macroarea)
summary(ejectives.null.model)

anova(ejectives.null.model, ejectives.elev.model1)
anova(ejectives.null.model, ejectives.elev.model2)
anova(ejectives.null.model, ejectives.elev.model3)
anova(ejectives.null.model, ejectives.elev.model4)
anova(ejectives.null.model, ejectives.elev.model5)


# ejectives.full.model <- glmer(data = final.data, glot_features_new ~ elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)
# 
# ejectives.reduced.model <- glmer(data = final.data, glot_features_new ~ mean_hum + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)
# anova(ejectives.reduced.model, ejectives.full.model)
```
### Tone

In a model comparison, humidity does seem to be a significant factor on tone ($\chi$^2^(1) = 2074.8, p \< 0.001).

I changed this to logistic (glmer).

```{r}
tone.hum.model <- glmer(data = final.data, tone_bin ~ mean_hum + (1 | stock) + (1 | macroarea), family = binomial)
summary(tone.hum.model)

tone.null.model <- glmer(data = final.data, tone_bin ~ (1 | stock) + (1 | macroarea), family = binomial)
anova(tone.null.model, tone.hum.model)

tone.full.model <- glmer(data = final.data, tone_bin ~ elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)

tone.reduced.model <- glmer(data = final.data, tone_bin ~ elev_binary + mean_temp + latitude_abs + diversity.new + (1 | stock), family = binomial)
anova(tone.reduced.model, tone.full.model)
```

### CV Ratio

Latitude is not significant ($\chi$^2^(1) = 0.5874, p = 0.443). 

```{r}
cv.hum.model <- lmer(data = final.data, cv_ratio_num ~ mean_hum + (1 | stock) + (1 | macroarea))
summary(cv.hum.model)

cv.null.model <- lmer(data = final.data, cv_ratio_num ~  (1 | stock) + (1 | macroarea))

anova(cv.null.model, cv.hum.model)

###later

cv.lat.model <- lmer(data = final.data, cv_ratio_num ~ latitude_abs + (1 | stock) + (1 | macroarea))
summary(cv.lat.model)
anova(cv.null.model, cv.lat.model)



cv.full.model <- lmer(data = final.data, cv_ratio_num ~ elev_binary + mean_hum + mean_temp + latitude_abs + diversity.new + (1 | stock))

cv.reduced.model <- lmer(data = final.data, cv_ratio_num ~ elev_binary + mean_hum + mean_temp + diversity.new + (1 | stock))
anova(cv.reduced.model, cv.full.model)



```

# Do not include

## First look to see whether GDP predicts features

### Ejectives

There is a marginal effect of GDP on whether or not a language has ejectives. 

```{r}
ejectives.GDP <- lm(data = final.data, glot_features_new ~ `2018`)
summary(ejectives.GDP)
```

### CV Ratio

There is a significant correlation between CV-ratio and GDP. 

```{r}
cv.GDP <- lm(data = final.data, cv_ratio_num ~ `2018`)
summary(cv.GDP)
```

### Tone

There is no significant correlation between GDP and tone. 

```{r}
tone.GDP <- lm(data = final.data, tone_num ~ `2018`)
summary(tone.GDP)

```


## GDP

### Altitude and GDP

There is a correlation between altitude and GDP

```{r}
elev.GDP <- lm(data = final.data, mean_hum ~ `2018`)
summary(elev.GDP)
```

### Latitude and GDP

There is a significant correlation between distance from the equator (absolute value of Latitude) and GDP.

```{r}
latitude.GDP <- lm(data = final.data, latitude_abs ~ `2018`)
summary(latitude.GDP)
```

### Humidity and GDP

There is a significant correlation between humidity and GDP. 

```{r}
hum.GDP <- lm(data = final.data, mean_hum ~ `2018`)
summary(hum.GDP)
```


## Test

```{r}
# dot plot = each dot = sub-famuiy
# x = average humidity
# y = proportion of languages in the subfamily with tone.communitysize
# color = language family

dot.plot.test <- final.data %>%
  #remove langs with no mbranch
  group_by(macroarea, area, family) %>%
  mutate(tone_bin = as.numeric(tone_bin)) %>%
  dplyr::summarise(mean_humidity = mean(mean_hum), n = n(), tone_sum = sum(tone_bin)) %>%
  # dplyr::summarise() %>%
  # dplyr::summarise() %>%
  mutate(tone_prop = tone_sum/n) 

                                
ggplot(dot.plot.test, aes(x = mean_humidity, y = tone_prop, size = n)) +
  geom_point() + 
  geom_smooth(method = "lm") +
  theme_bw() +
  labs(x = "Mean Humidity",
       y = "Proportion of Languages with tone in Language Family")

#exclude families with only 1 langauge
dot.plot.test2 <- dot.plot.test %>%
  filter(n != 1)

ggplot(dot.plot.test2, aes(x = mean_humidity, y = tone_prop, size = n, color = area)) +
  geom_point()+ 
  theme_bw() +
  labs(x = "Mean Humidity",
       y = "Proportion of Languages with tone in Language Family", 
       title = "Exclude Language families with only 1 language") +
  facet_wrap(~area)

```
